requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-currentuser==0.5.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
